Update setup.cfg
change of name to duhast to be compatible with PEP 625 :

This email is notifying you of an upcoming deprecation that we have determined may affect you as a result of your recent upload to 'DuHast'.

In the future, PyPI will require all newly uploaded source distribution filenames to comply with PEP 625. Any source distributions already uploaded will remain in place as-is and do not need to be updated.

Specifically, your recent upload of 'DuHast-1.1.8.tar.gz' is incompatible with PEP 625 because it does not contain the normalized project name 'duhast'.

In most cases, this can be resolved by upgrading the version of your build tooling to a later version that supports PEP 625 and produces compliant filenames.

If you have questions, you can email admin@pypi.org to communicate with the PyPI admin@pypi.org to communicate with the PyPI administrator